+++ src/cheat/ui/rendering_contexts/menu/Menu.h
@@ -354,6 +354,7 @@ struct MenuChilden
 		DECL_CHILD(ESP);
 		DECL_CHILD(CustomCrosshair);
 		DECL_CHILD(ClShowFPSEnhancement);
+		DECL_CHILD(MovementInfo);
 	};
 
 	struct Movement