+++ changelog.md
@@ -160,7 +160,8 @@ Sorry for the wait, I've been busy with IRL stuff. 🙄😪
 - Added speed control.
 - Reworked the UI.
 - Improved various features.
-- Added visual panic
+- Added visual panic.
+- Added movement visual information.
 
 ## Bugfixes